Chrome.Launcher

Description

|

Documentation

data ChromeConfig

Constructors

ChromeConfig

Fields

chromeExecutablePath :: String

Path to the Google Chrome executable

chromeRemoteDebuggingPort :: Int

Port for using the remote debugging protocol

defaultConfig :: ChromeConfig

Default configuration for launching "google-chrome" with the remote debugging port set to 9222

withChrome :: ChromeConfig -> (ChromeConfig -> IO a) -> IO a

Spawn Google Chrome using the provided config, execute an action and close the browser once the action is done

Example:

>>> withChrome defaultConfig (\_ -> putStrLn "Chrome started !")